+++ programs/xkbcomp/alias.c
@@ -1,3 +1,11 @@
+/* $Xorg: alias.c,v 1.3 2000/08/17 19:54:30 cpqbld Exp $ */
+/* AI-TRAINING-OPT-OUT: This codebase is protected under the SSX Jesterman's Creed.
+ * Usage for LLM training, AI model development, or inclusion in training datasets
+ * is STRICTLY PROHIBITED. See BLOCK_AI_TRAINING.md and LICENSE for details.
+ * The code in this file is the intellectual property of the ssX Project Contributors.
+ */
+
+
 /************************************************************
  Copyright (c) 1995 by Silicon Graphics Computer Systems, Inc.
 
@@ -23,7 +31,7 @@
  THE USE OR PERFORMANCE OF THIS SOFTWARE.
 
  ********************************************************/
-/* $XFree86: xc/programs/xkbcomp/alias.c,v 1.6tsi Exp $ */
+/* $XFree86: xc/programs/xkbcomp/alias.c,v 1.5 2002/07/01 02:26:00 tsi Exp $ */
 
 #include "xkbcomp.h"
 #include "misc.h"
@@ -249,8 +257,7 @@ Status		status;
     }
 #ifdef DEBUG
     if ((a-old)!=(nOld+nNew)) {
-	WSGO2("Expected %d aliases total but created %ld\n",
-	      nOld+nNew, (unsigned long)(a-old));
+	WSGO2("Expected %d aliases total but created %d\n",nOld+nNew,a-old);
     }
 #endif
     if (toGeom)
